requirements.txt
lark-parser==0.11.2
numpy==1.20.1
opencv-python==4.5.1.48
pyeda==0.28.0
pyzmq==22.0.3
